refactor(info) get npm package method (#3045)
